post_docking_analysis/requirements.txt
# Post-Docking Analysis Requirements

# Core dependencies
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
pyyaml>=6.0

# Optional dependencies
openbabel>=3.1.1  # For pose extraction
pymol>=2.5.0      # For 3D visualizations
pandamap>=1.0.0   # For interaction analysis

# For Excel output support
openpyxl>=3.0.0

# For advanced visualization
plotly>=5.0.0